refactor<> crud 관련 조회 로직만 다루는 서비스 로직 분리
추상화 레벨이 높아지는거 같긴한데.. 추후 확장성을 고려하여 분리하였습니다.
