+++ src/main/java/com/gulbi/Backend/domain/rental/product/service/category/CategoryCrudService.java
@@ -1,6 +1,6 @@
 package com.gulbi.Backend.domain.rental.product.service.category;
 
-import com.gulbi.Backend.domain.rental.product.dto.CategoryProjection;
+import com.gulbi.Backend.domain.rental.product.dto.category.CategoryProjection;
 import com.gulbi.Backend.domain.rental.product.entity.Category;
 
 import java.util.List;
